fix(tightfollow2pnw): use capnp enum comparison, not int(), for personality check
int(personality) crashed plannerd on every tick with TypeError: int() argument
must be a string, a bytes-like object or a real number, not
'capnp.lib.capnp._DynamicEnum' -- pycapnp enum readers don't support int().
The existing long_mpc.py get_T_FOLLOW/get_jerk_factor already prove direct
== / != comparison against log.LongitudinalPersonality.X works correctly on
this exact object (same sm['selfdriveState'].personality reader). Confirmed
by independent Opus review: fix is correct, no sibling bugs in the rest of
the tightfollow2pnw/speedlimitdebug2pnw diff.

This was live on the car for several minutes and took plannerd down entirely.

+++ selfdrive/controls/lib/pnw_vehicle.py
@@ -355,6 +355,6 @@ class PnwVehicle:
     t_follow_override; None there means "use the shared get_T_FOLLOW(personality) as before"."""
     if not self.tight_aggressive_follow:
       return None
-    if int(personality) != int(log.LongitudinalPersonality.aggressive):
+    if personality != log.LongitudinalPersonality.aggressive:
       return None
     return _TIGHT_AGGRESSIVE_T_FOLLOW
